rest controller for submissions
controller and assignment dto

+++ src/main/java/edu/depaul/cdm/se452/d2l_mock/submission/AssignmentRepository.java
@@ -3,6 +3,6 @@ package edu.depaul.cdm.se452.d2l_mock.submission;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface AssignmentRepository extends JpaRepository<Assignment, Integer>{
+public interface AssignmentRepository extends JpaRepository<Assignment, Long>{
 
 }
